Fix tr/// empty replacement and big-endian integer pack/unpack bugs
Two critical bug fixes for PerlOnJava pack/unpack functionality:

1. tr/// Empty Replacement Bug (RuntimeTransliterate.java):
   - Fixed transliteration with empty replacement to count matches instead of deleting
   - Changed logic to map characters to themselves (identity mapping) when replacement is empty
   - Impact: +12 tests passing

2. Big-Endian Integer Pack/Unpack Bug:
   - Fixed NumericPackHandler.java: i/I formats now respect endianness modifiers (>, <)
   - Fixed UnpackState.java: Added currentByteOrder field to persist byte order across buffer recreations
   - Fixed NumericFormatHandler.java: Changed int to long to avoid sign extension issues
   - Root cause: i/I formats were hardcoded to little-endian, ignoring modifiers
   - Impact: +200-500 tests (affects checksums, cross-platform compatibility)

Test Results:
- pack.t: 9,438 / 14,724 passing (64% pass rate)
- Total improvement: +501 tests (+5.6% from baseline)

Files Modified:
- src/main/java/org/perlonjava/operators/RuntimeTransliterate.java
- src/main/java/org/perlonjava/operators/pack/NumericPackHandler.java
- src/main/java/org/perlonjava/operators/UnpackState.java
- src/main/java/org/perlonjava/operators/unpack/NumericFormatHandler.java
- dev/prompts/fix-pack-test-failures.md

+++ src/main/java/org/perlonjava/operators/unpack/NumericFormatHandler.java
@@ -79,9 +79,10 @@ public abstract class NumericFormatHandler implements FormatHandler {
                 if (buffer.remaining() < 4) {
                     break;
                 }
-                int value = buffer.getInt();
+                // Read as long to avoid sign extension issues with unsigned values
+                long value = buffer.getInt();
                 if (signed) {
-                    output.add(new RuntimeScalar(value));
+                    output.add(new RuntimeScalar((int) value));
                 } else {
                     output.add(new RuntimeScalar(value & 0xFFFFFFFFL));
                 }
